+++ ConsoleSand/grid/grid.hpp
@@ -9,13 +9,13 @@
 class Grid
 {
 public:
-	Grid(const uint8_t width, const uint8_t height);
+	Grid(const uint16_t width, const uint16_t height);
 	Grid(Grid& otherGrid);
 	Grid& operator= (Grid& otherGrid);
 
 	~Grid();
 
-	std::pair<uint8_t, uint8_t> size() const;
+	std::pair<uint16_t, uint16_t> size() const;
 
 	struct Cell
 	{
@@ -33,33 +33,33 @@ public:
 
 	void process();
 
-	bool spawnMaterial(const uint8_t& x, const uint8_t& y, const Materials::Id& mId);
+	bool spawnMaterial(const uint16_t& x, const uint16_t& y, const Materials::Id& mId);
 
 	inline void clearAll();
 
 private:
 
-	inline Cell** allocate(uint8_t width, uint8_t height);
-	inline void deallocate(Cell** cells, uint8_t width);
+	inline Cell** allocate(uint16_t width, uint16_t height);
+	inline void deallocate(Cell** cells, uint16_t width);
 	inline void copyCells(Cell** const cells);
 
 	void inline clearMoveState();
 
 	typedef enum class GridBorder { Left = 0, Upper, Right, Bottom } gridBorderSpecify;
-	bool inline trespassing(const int8_t& val, const gridBorderSpecify& whatBorder) const;
+	bool inline trespassing(const int16_t& val, const gridBorderSpecify& whatBorder) const;
 
 	void inline swapCells(Cell& cell, Cell& newCell);
 	bool inline clearCell(Cell& cell);
 
-	bool processAcidic(Cell& cell, const int8_t& x, const int8_t& y);
-	bool processFlamable(Cell& cell, const int8_t& x, const int8_t& y);
-	bool processDiffusing(Cell& cell, const int8_t& x, const int8_t& y);
+	bool processAcidic(Cell& cell, const int16_t& x, const int16_t& y);
+	bool processFlamable(Cell& cell, const int16_t& x, const int16_t& y);
+	bool processDiffusing(Cell& cell, const int16_t& x, const int16_t& y);
 
-	bool processPowdery(Cell& cell, const int8_t& x, const int8_t& y);
-	bool processLiquid(Cell& cell, const int8_t& x, const int8_t& y);
-	bool processGas(Cell& cell, const int8_t& x, const int8_t& y);
+	bool processPowdery(Cell& cell, const int16_t& x, const int16_t& y);
+	bool processLiquid(Cell& cell, const int16_t& x, const int16_t& y);
+	bool processGas(Cell& cell, const int16_t& x, const int16_t& y);
 
-	uint8_t m_width;
-	uint8_t m_height;
+	uint16_t m_width;
+	uint16_t m_height;
 	Cell** m_grid;
 };
